+++ src/components/user-item/index.tsx
@@ -1,4 +1,4 @@
-import { Text, View } from 'react-native';
+import { Text } from 'react-native';
 import { WrapperCardUserList } from './styles';
 interface UserProps {
   id: string;
